feat(swiftui home): some refactors and removing unnecessary protocols
